Add spillDuration support

+++ src/main/java/analyzer/SlowestTaskComparator.java
@@ -563,7 +563,7 @@ public class SlowestTaskComparator {
         for (String line : excutorLogLines) {
             if (line.contains("[Task " + taskId + " SpillMetrics]")) {
                 int writeTime = Integer.parseInt(line.substring(line.indexOf("writeTime") + 12, line.indexOf(" s,")));
-                long recordsWritten = Integer.parseInt(line.substring(line.indexOf("recordsWritten") + 17, line.indexOf(", bytesWritten")));
+                long recordsWritten = Long.parseLong(line.substring(line.indexOf("recordsWritten") + 17, line.indexOf(", bytesWritten")));
                 spillTime += writeTime;
             }
         }
